requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
apache-airflow==2.9.1
requests==2.32.3
pytest==8.2.2
python-dotenv==1.0.1
